refactor(#316): extract resolvedNodeOpts + loadHandicap to campaignHelpers; extract usePlaytime hook
Moves two categories of logic out of App.tsx (-130 lines):

1. Pure game helpers → web/src/game/campaignHelpers.ts
   - resolvedNodeOpts: computes difficulty-scaled NewGameOptions from a
     QuestNode, run count, and active modifiers
   - loadHandicap/HANDICAP_KEY: localStorage handicap accessor

2. usePlaytime hook → web/src/hooks/usePlaytime.ts
   - Encapsulates gameTimerRef/battleTimerRef, flushPlaytimeToStorage,
     applyPlaytimeTransition, and the two playtime useEffects
   - App.tsx calls usePlaytime({ screen, isTabHidden, setAchievementToasts })
     and destructures { flushPlaytimeToStorage }
